requirements.txt
requests>=2.31.0,<3.0.0
pypdf>=6.9.0,<7.0.0
sec-edgar-downloader>=5.0.2,<6.0.0
psycopg2-binary>=2.9.9,<3.0.0
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
